+++ dd-java-agent/agent-profiling/profiling-controller/src/main/java/com/datadog/profiling/controller/ControllerContext.java
@@ -4,6 +4,12 @@ import com.datadog.profiling.utils.ProfilingMode;
 import java.util.EnumSet;
 import java.util.Set;
 
+/**
+ * The controller context is for accumulating information about controllers which provides a simple
+ * protocol for the controllers to communicate with each other. This is necessary because the
+ * controllers are not actually self-contained, e.g. some JFR events are mutually exclusive with the
+ * data collected by DDProf.
+ */
 public class ControllerContext {
 
   private boolean isDatadogProfilerActive;
@@ -25,6 +31,11 @@ public class ControllerContext {
     this.datadogProfilingModes = datadogProfilingModes;
   }
 
+  /**
+   * A snapshot is an immutable copy of the context (state shared between controllers)
+   *
+   * @return an immutable snapshot of the context
+   */
   public Snapshot snapshot() {
     return new Snapshot(this);
   }
